fixed helpers
