--- src/components/main/actionsPanel/navigatorPanel/helpers/getFileNameFromPath.ts
@@ -1,5 +1,0 @@
-import { TFile } from "@_types/main";
-
-export const getFileNameFromPath = (file: TFile) => {
-  return file.uid.split("/")[file.uid.split("/").length - 1];
-};
