+++ src/components/main/actionsPanel/navigatorPanel/helpers/index.ts
@@ -1,5 +1,1 @@
-export { isSelected } from "./isSelected";
-export { isHomeIcon } from "./isHomeIcon";
-export { getFileNameFromPath } from "./getFileNameFromPath";
-export { setWorkspaceFavicon } from "./setWorkspaceFavicon";
-export { selectFirstNode } from "./selectFirstNode";
+export * from "./helpers";
